Added GDB source to eclipse build on Linux, and fixed some warnings

+++ iss/src/rv32_cpu_gdb.h
@@ -61,7 +61,7 @@ typedef SOCKET rv32gdb_skt_t;
 // -------------------------------------------------------------------------
 
 // Map the socket type for Linux  
-typedef long rv32gdb_skt_t;
+typedef long long rv32gdb_skt_t;
 
 // -------------------------------------------------------------------------
 // DEFINITIONS (Linux)
